fix: wire public speech CORS through DI
